requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
aiofiles==0.8.0
    # via sanic
aiopg==1.3.5
    # via -r requirements/requirements.in
alembic==1.8.1
    # via -r requirements/requirements.in
arrow==1.2.3
    # via -r requirements/requirements.in
async-timeout==4.0.2
    # via aiopg
click==8.1.3
    # via -r requirements/requirements.in
httptools==0.3.0
    # via sanic
mako==1.1.6
    # via alembic
markupsafe==2.0.1
    # via mako
marshmallow==3.18.0
    # via -r requirements/requirements.in
multidict==5.2.0
    # via sanic
packaging==21.3
    # via
    #   marshmallow
    #   sanic-cors
psycopg2-binary==2.9.3
    # via aiopg
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via arrow
python-dotenv==0.21.0
    # via -r requirements/requirements.in
pyyaml==6.0
    # via -r requirements/requirements.in
sanic==22.9.0
    # via
    #   -r requirements/requirements.in
    #   sanic-cors
sanic-cors==2.2.0
    # via -r requirements/requirements.in
sanic-routing==22.8.0
    # via sanic
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.41
    # via
    #   -r requirements/requirements.in
    #   alembic
ujson==5.1.0
    # via sanic
uvloop==0.16.0
    # via sanic
websockets==10.1
    # via sanic
